feat(scroll-to-top): Add scroll-to-top button component
