+++ src/components/scroll-to-top/ScrollToTop.module.css
@@ -1,0 +1,23 @@
+& .scrollToTopButton {
+  position: fixed;
+  bottom: 20px;
+  right: 20px;
+  background: var(--gray-2);
+  color: var(--gray-9);
+  border-radius: 5px;
+  border: none;
+  padding: 8px 12px;
+  font-size: 16px;
+  cursor: pointer;
+  transition: background .2s;
+  z-index: 10;
+  width: 40px;
+  height: 40px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+
+  &:hover {
+    background: var(--gray-3);
+  }
+}
